tests added

+++ homework_01/task_02/src/utils.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "utils.cpp"
+// #include "utils.cpp"
 #include <string>
 #include <string_view>
 #include <vector>
@@ -8,4 +8,4 @@
 std::string CalcTwo(std::vector<std::string> vec);
 std::vector<std::string> SplitToTokens(const std::string &data);
 std::vector<std::string> Calc(std::vector<std::string> vec);
-int Calculate(const std::string &data);
+double Calculate(const std::string &data);
